Abstract the IO verifier for use in other test files
